+++ docs/md/Version-history.md
@@ -8,6 +8,7 @@
 
 Changes in [pre-release builds](https://www.sumatrapdfreader.org/prerelease):
 
+- in Command Palette, if you start search with ":" we show everything (like in 3.5)
 - in Command Palette, when viewing opened files history (#), you can press Delete to remove the entry from history
 - improved zooming:
   - zooming with pinch touch screen gesture or with ctrl + scroll wheel now zooms around the mouse position and does continuous zoom levels. Used to zoom around top-left corner and progress fixed zoom levels shown in menu